[chore] : move hane class to hane directory
