+++ src/test/java/kr/where/backend/member/memberServiceTest.java
@@ -1,7 +1,7 @@
 package kr.where.backend.member;
 
 import kr.where.backend.api.json.CadetPrivacy;
-import kr.where.backend.api.json.Hane;
+import kr.where.backend.api.json.hane.Hane;
 import kr.where.backend.auth.authUser.AuthUser;
 import kr.where.backend.group.GroupMemberRepository;
 import kr.where.backend.group.GroupRepository;
